+++ packages/client/src/components/mfm.ts
@@ -47,7 +47,7 @@ export default defineComponent({
 	render() {
 		if (this.text == null || this.text === '') return;
 
-		const ast = (this.plain ? mfm.parsePlain : mfm.parse)(this.text, { fnNameList: MFM_TAGS });
+		const ast = (this.plain ? mfm.parseSimple : mfm.parse)(this.text, { fnNameList: MFM_TAGS });
 
 		const validTime = (t: string | null | undefined) => {
 			if (t == null) return null;